requirements.txt
deta==1.1.0
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
Jinja2==3.1.1
psycopg2-binary==2.9.5
python-multipart==0.0.5
SQLAlchemy==1.4.39
pandas==1.5.2
